refactor: イベント機能をfeatureディレクトリに集約
