+++ app/web/app/.server/actions/index.ts
@@ -1,4 +1,3 @@
-export * from "./eventAction";
 export * from "./friendAction";
 export * from "./memberAction";
 export * from "./paymentAction";
