+++ app/web/app/.server/loaders/index.ts
@@ -1,5 +1,3 @@
-export * from "./eventLoader";
-export * from "./eventsLoader";
 export * from "./friendsLoader";
 export * from "./membersLoader";
 export * from "./paymentsLoader";
