+++ app/web/app/features/events/components/create-event-form.tsx
@@ -6,13 +6,14 @@ import {
 	useInputControl,
 } from "@conform-to/react";
 import { parseWithZod } from "@conform-to/zod";
-import { CalendarIcon } from "@radix-ui/react-icons";
-import { PopoverClose } from "@radix-ui/react-popover";
 import { Form, useActionData, useNavigation } from "@remix-run/react";
 import { format } from "date-fns";
 import { useId } from "react";
-import type { EventAction } from "~/.server/actions";
 import type { EventEndpoints_EventPostRequest, Member } from "~/api/@types";
+import { cn } from "~/lib/utils";
+
+import { CalendarIcon } from "@radix-ui/react-icons";
+import { PopoverClose } from "@radix-ui/react-popover";
 import { Button } from "~/components/ui/button";
 import { Calendar } from "~/components/ui/calendar";
 import { Input } from "~/components/ui/input";
@@ -29,23 +30,24 @@ import {
 	SelectTrigger,
 	SelectValue,
 } from "~/components/ui/select";
-import { cn } from "~/lib/utils";
-import { eventCreateFormSchema } from "~/schema/eventFormSchema";
+
+import type { CreateEventAction } from "../actions";
+import { createEventSchema as schema } from "../schemas";
 
 interface Props {
 	defaultValue?: Partial<EventEndpoints_EventPostRequest>;
 	members: Member[];
 }
 
-export function EventCreateForm({ defaultValue, members }: Props) {
-	const actionData = useActionData<EventAction>();
+export function CreateEventForm({ defaultValue, members }: Props) {
+	const actionData = useActionData<CreateEventAction>();
 
 	const [form, { name, eventedAt, amount, payments, paidBy }] = useForm({
 		defaultValue,
 		lastResult: actionData?.submission,
 		onValidate({ formData }) {
 			return parseWithZod(formData, {
-				schema: eventCreateFormSchema,
+				schema,
 			});
 		},
 	});
