+++ app/web/app/features/events/components/index.ts
@@ -1,0 +1,4 @@
+export * from "./create-event-form";
+export * from "./event-card";
+export * from "./event-list";
+export * from "./update-event-form";
