+++ app/web/app/features/events/loaders/index.ts
@@ -1,0 +1,2 @@
+export * from "./event-list-loader";
+export * from "./event-loader";
