+++ app/web/app/routes/_auth.groups_.$groupId_.events_.$eventId.tsx
@@ -1,32 +1,25 @@
 import type { MetaFunction } from "@remix-run/cloudflare";
 import { Await, useActionData, useLoaderData } from "@remix-run/react";
 import { Suspense, useEffect } from "react";
-import type { EventAction } from "~/.server/actions";
-import type { EventLoader } from "~/.server/loaders";
-import { EventUpdateForm } from "~/components/EventUpdateForm";
+import { Spinner } from "~/components";
 import { useToast } from "~/components/ui/use-toast";
 
-export { eventAction as action } from "~/.server/actions";
-export { eventLoader as loader } from "~/.server/loaders";
+import type { UpdateEventAction } from "~/features/events/actions";
+import { UpdateEventForm } from "~/features/events/components";
+import type { EventLoader } from "~/features/events/loaders";
+export { createEventAction as action } from "~/features/events/actions";
+export { eventLoader as loader } from "~/features/events/loaders";
 
 export const meta: MetaFunction = () => [
 	{ title: "Datti | イベント編集" },
 	{ name: "description", content: "誰にいくら払ったっけ？を記録するサービス" },
 ];
 
-function LoadingSpinner() {
-	return (
-		<div className="w-full min-h-[60vh] grid place-content-center">
-			<div className="animate-spin h-10 w-10 border-4 border-blue-500 rounded-full border-t-transparent" />
-		</div>
-	);
-}
-
 export default function EventDetail() {
 	const { toast } = useToast();
 
 	const { event, members } = useLoaderData<EventLoader>();
-	const actionData = useActionData<EventAction>();
+	const actionData = useActionData<UpdateEventAction>();
 	useEffect(() => {
 		if (actionData) {
 			toast({
@@ -42,12 +35,12 @@ export default function EventDetail() {
 			</div>
 			<div className="rounded-lg bg-white py-3 px-5">
 				<div className="flex flex-col py-3 gap-7">
-					<Suspense fallback={<LoadingSpinner />}>
+					<Suspense fallback={<Spinner />}>
 						<Await resolve={event}>
 							{(event) => (
 								<Await resolve={members}>
 									{({ members }) => (
-										<EventUpdateForm defaultValue={event} members={members} />
+										<UpdateEventForm defaultValue={event} members={members} />
 									)}
 								</Await>
 							)}
